third_party/blink/web_tests/external/wpt/css/css-shadow/slotted-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] "::slotted(*):is()" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):is()' is not a valid selector.
[FAIL] "::slotted(*):is(:hover)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):is(:hover)' is not a valid selector.
[FAIL] "::slotted(*):is(#id)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):is(#id)' is not a valid selector.
[FAIL] "::slotted(*):where()" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):where()' is not a valid selector.
[FAIL] "::slotted(*):where(:hover)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):where(:hover)' is not a valid selector.
[FAIL] "::slotted(*):where(#id)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):where(#id)' is not a valid selector.
[FAIL] "::slotted(*):where(::before)" should be a valid selector
  Failed to execute 'querySelector' on 'Document': '::slotted(*):where(::before)' is not a valid selector.
[FAIL] "::slotted(select)::picker(select)" should be a valid selector
  assert_equals: serialization should be canonical expected "::slotted(select)::picker(select)" but got "::slotted(select)::picker"
Harness: the test ran to completion.
